pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "systemeval"
version = "0.4.0"
description = "AI-powered test orchestration platform: traditional testing (pytest, jest, playwright) + DebuggAI natural language tests, commit-based generation, and visual failure analysis"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "DebuggAI", email = "dev@debugg.ai"}
]
keywords = ["testing", "pytest", "jest", "cli", "test-runner", "pipeline", "e2e"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
pytest = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",
]
jest = []
# Django support for pipeline adapter and Django project testing
# Install with: pip install systemeval[django]
django = [
    "Django>=4.0",
]
# Pipeline adapter - requires Django for database metrics collection
# In container environments, Django is typically already available
# Install with: pip install systemeval[pipeline] or systemeval[django]
pipeline = [
    "systemeval[django]",
]
dev = [
    "systemeval[pytest]",
    "black",
    "ruff",
    "mypy",
]

[project.scripts]
systemeval = "systemeval.cli:main"

[project.urls]
Homepage = "https://debugg.ai"
Documentation = "https://debugg.ai/docs/systemeval"
Repository = "https://github.com/debugg-ai/systemeval"
Issues = "https://github.com/debugg-ai/systemeval/issues"
Changelog = "https://github.com/debugg-ai/systemeval/blob/main/CHANGES.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["systemeval*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

# =============================================================================
# SE-FLW: Architecture Lint Rules
# =============================================================================
# Custom lint checks are in scripts/check_env_access.py
#
# Rule: No direct environment variable access in business logic
# Principle: "No magic values" and "no config discovery"
#
# Forbidden patterns (in business logic):
#   - os.environ["KEY"]
#   - os.environ.get("KEY")
#   - os.getenv("KEY")
#   - environ.get patterns
#
# Allowed files (entry points and config layer):
#   - cli.py (entry point)
#   - config.py (config loading layer)
#   - conftest.py (test setup)
#   - tests/* (test files)
#
# Allowed patterns (subprocess env passthrough):
#   - env={**os.environ, ...}  (passing env to subprocess)
#   - env=dict(os.environ)     (copying env for subprocess)
#
# Run check: python scripts/check_env_access.py
#
# -----------------------------------------------------------------------------
# SE-7gk: No hardcoded paths in business logic
# -----------------------------------------------------------------------------
# Rule: Paths must come from configuration, not be hardcoded
# Principle: "No magic values" - paths are environment-specific
#
# Forbidden patterns (in business logic):
#   - /Users/*, /home/*, /root/*  (user directories)
#   - C:\*, D:\*                   (Windows drives)
#   - /var/*, /tmp/*, /opt/*       (system directories)
#   - /Applications/*, /Library/*  (macOS system paths)
#
# Allowed files (configuration layer):
#   - config.py, settings.py, constants.py, paths.py, defaults.py
#   - conftest.py (test setup)
#   - tests/* (test files)
#   - scripts/* (lint scripts)
#
# Allowed contexts:
#   - Docstrings (documentation examples are okay)
#   - Comments
#
# Run check: python scripts/check_hardcoded_paths.py
# =============================================================================

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Run pytest adapter tests in isolation to avoid module caching conflicts
addopts = "-p no:django --ignore=tests/test_pytest_adapter.py"
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning",
]

# Note: Run test_pytest_adapter.py separately:
# python3 -m pytest tests/test_pytest_adapter.py -p no:django
